Pricing Fixed Income + Saving in the entity Trade Signed-off-by: Portable DELL <Portable DELL@172.19.10.170>

+++ ideca2/ideca2-client/src/main/java/pdev/financialbrains/client/GUI/BackOfficeUser/PricingCorporateBond.java
@@ -122,7 +122,7 @@ public class PricingCorporateBond extends JFrame {
 				Integer Months = Integer.parseInt(tf_months.getText());
 
 				
-				Double Price=FixedIncomeSecurityManagementDelegate.doPriceCorpBond(FaceValue, TimeToMaturity,InterestRate , Frequency, CouponRate, Months);
+				Float Price=FixedIncomeSecurityManagementDelegate.doPriceCorpBond(FaceValue, TimeToMaturity,InterestRate , Frequency, CouponRate, Months);
 
 				tf_price.setText(Price.toString());
 			}
